initial commit fling project.
I want an easier way to stack up a bunch of scripts and tools and have simple autocomplete for them... even if they're in a bunch of different binaries, languages, *whatever*.

And it should be copy-pasta-able bash.

Because why shouldn't it be?

Ideally, this will be able to replace common uses of "make".  It won't do dependency resolve, but it *will* be a lot better than coding a bunch of opt parsing by hand, and can be a conventional entrypoint to all sorts of projects.

Signed-off-by: Eric Myhre <hash@exultant.us>